Stuff optimized, 2.5 speedup

+++ Tank.cpp
@@ -13,7 +13,7 @@ Tank::Tank(
     Sprite* smoke_sprite,
     float tar_x,
     float tar_y,
-    float collision_radius,
+    UINT16 collision_radius,
     int health,
     float max_speed)
     : position(pos_x, pos_y),
